Moved logic from componenet to container
