+++ src/components/Sidebar.js
@@ -1,42 +1,10 @@
 import React, { PropTypes } from 'react'
-import { connect } from 'react-redux';
-import { ARTICLE, VIDEO, LIVE } from '../constants/ArticleTypes';
+
 import '../css/Sidebar.css';
 import Icon from '../components/Icon';
 import UnUsedContentContainer from '../containers/UnUsedContentContainer';
 
-const mapStateToProps = ({articles}) => {
-  const content = {};
-  const contentTypes = [ARTICLE, VIDEO, LIVE];
 
-  contentTypes.forEach( type => {
-    content[type] = Object.keys(articles)
-      .filter(key => articles[key].type === type)
-      .map(key => articles[key]);
-  });
-
-  return {
-    contentTypes,
-    content
-  };
-}
-
-const mapDispatchToProps = () => {
-  return {
-    toggleSidebar: () => {
-      const sidebarClasses = document.getElementById("sidebar").classList;
-      if(sidebarClasses.contains("Sidebar-wrapper-active")){
-        sidebarClasses.remove("Sidebar-wrapper-active");
-        sidebarClasses.add("Sidebar-wrapper-inactive");
-      }
-      else{
-        sidebarClasses.remove("Sidebar-wrapper-inactive");
-        sidebarClasses.add("Sidebar-wrapper-active");
-      }
-
-    }
-  }
-}
 
 const Sidebar = ({contentTypes, content, toggleSidebar}) => (
   <div className="Sidebar-wrapper" id="sidebar">
@@ -66,4 +34,4 @@ Sidebar.propTypes = {
   content: PropTypes.object.isRequired
 }
 
-module.exports = connect(mapStateToProps, mapDispatchToProps)(Sidebar);
+module.exports = Sidebar;
