--- src/containers/SidebaContainer.js
@@ -1,1 +1,0 @@
-import Siderbar from '../components/Siderbar';
